requirements.txt
# PyTorch avec support MPS pour M2 Max
torch>=2.0.0
torchvision>=0.15.0

# Diffusers et transformers
diffusers>=0.24.0
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.7.0

# ControlNet et utilitaires image
controlnet-aux>=0.0.7
# xformers - RETIRÉ (problème compilation macOS)
safetensors>=0.4.0
pillow>=9.5.0
opencv-python>=4.8.0
numpy>=1.24.0

# Interface web
streamlit>=1.28.0

# API (optionnel)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# AWS et utilitaires
boto3>=1.29.0
python-dotenv>=1.0.0

# Dépendances additionnelles pour macOS
scipy>=1.10.0
protobuf>=3.20.0

# Optionnel pour réduire les warnings
mediapipe>=0.10.0